Maze final update before presentation
