+++ Maze_Solver_Source_Code/input.txt
@@ -1,0 +1,7 @@
+4
+3
+2
+1,1
+2,2
+0,0
+3,2
