+++ Maze_Solver_Source_Code/input2.txt
@@ -1,0 +1,26 @@
+10
+10
+21
+1,3
+2,3
+3,3
+4,2
+4,3
+4,4
+5,1
+5,2
+5,3
+6,2
+6,3
+6,4
+7,3
+7,4
+7,5
+8,4
+8,5
+8,6
+9,5
+9,6
+9,7
+0,0
+9,9
