+++ Maze_Solver_Source_Code/orginal.txt
@@ -1,0 +1,25 @@
+10
+10
+20
+2,8
+3,8
+5,7
+6,7
+8,7
+3,6
+4,6
+8,5
+1,4
+3,4
+5,4
+7,4
+8,4
+3,3
+1,2
+6,2
+3,1
+5,1
+6,1
+8,1
+4,8
+8,0
